Move base diff function
