+++ edsl/questions/__init__.py
@@ -3,26 +3,26 @@ from edsl.questions.settings import Settings
 from edsl.questions.register_questions_meta import RegisterQuestionsMeta
 
 # Base Class
-from edsl.questions.QuestionBase import QuestionBase
+from .QuestionBase import QuestionBase
 
 # Core Questions
-from edsl.questions.QuestionCheckBox import QuestionCheckBox
-from edsl.questions.QuestionExtract import QuestionExtract
-from edsl.questions.QuestionFreeText import QuestionFreeText
-from edsl.questions.QuestionFunctional import QuestionFunctional
-from edsl.questions.QuestionList import QuestionList
-from edsl.questions.QuestionMatrix import QuestionMatrix
-from edsl.questions.QuestionDict import QuestionDict
-from edsl.questions.QuestionMultipleChoice import QuestionMultipleChoice
-from edsl.questions.QuestionNumerical import QuestionNumerical
-from edsl.questions.QuestionBudget import QuestionBudget
-from edsl.questions.QuestionRank import QuestionRank
+from .QuestionCheckBox import QuestionCheckBox
+from .QuestionExtract import QuestionExtract
+from .QuestionFreeText import QuestionFreeText
+from .QuestionFunctional import QuestionFunctional
+from .QuestionList import QuestionList
+from .QuestionMatrix import QuestionMatrix
+from .QuestionDict import QuestionDict
+from .QuestionMultipleChoice import QuestionMultipleChoice
+from .QuestionNumerical import QuestionNumerical
+from .QuestionBudget import QuestionBudget
+from .QuestionRank import QuestionRank
 
 # # # Questions derived from core questions
-from edsl.questions.derived.QuestionLikertFive import QuestionLikertFive
-from edsl.questions.derived.QuestionLinearScale import QuestionLinearScale
-from edsl.questions.derived.QuestionYesNo import QuestionYesNo
-from edsl.questions.derived.QuestionTopK import QuestionTopK
+from .derived.QuestionLikertFive import QuestionLikertFive
+from .derived.QuestionLinearScale import QuestionLinearScale
+from .derived.QuestionYesNo import QuestionYesNo
+from .derived.QuestionTopK import QuestionTopK
 
 # # Compose Questions
 # from edsl.questions.compose_questions import compose_questions
